+++ src/serial_clear_buffer_in.cpp
@@ -9,14 +9,15 @@ extern "C"
 
     MODULE_API auto serialClearBufferIn(int64_t handle, ErrorCallbackT error_callback) -> int
     {
-        cpp_bindings_linux::detail::HandleContext context;
-        const auto rc = cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &context);
-        if (rc < 0)
+        cpp_bindings_linux::detail::HandleContext handle_context;
+        const auto status =
+            cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &handle_context);
+        if (status < 0)
         {
-            return rc;
+            return status;
         }
 
-        if (tcflush(context.fd, TCIFLUSH) != 0)
+        if (tcflush(handle_context.file_descriptor, TCIFLUSH) != 0)
         {
             return cpp_bindings_linux::detail::failErrno<int>(
                 error_callback, cpp_bindings_linux::detail::statusValue(cpp_core::StatusCode::Io::kClearBufferInError));
