+++ src/serial_clear_buffer_out.cpp
@@ -9,20 +9,21 @@ extern "C"
 
     MODULE_API auto serialClearBufferOut(int64_t handle, ErrorCallbackT error_callback) -> int
     {
-        cpp_bindings_linux::detail::HandleContext context;
-        const auto rc = cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &context);
-        if (rc < 0)
+        cpp_bindings_linux::detail::HandleContext handle_context;
+        const auto status =
+            cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &handle_context);
+        if (status < 0)
         {
-            return rc;
+            return status;
         }
 
-        if (tcdrain(context.fd) != 0)
+        if (tcdrain(handle_context.file_descriptor) != 0)
         {
             return cpp_bindings_linux::detail::failErrno<int>(
                 error_callback, cpp_bindings_linux::detail::statusValue(cpp_core::StatusCode::Io::kClearBufferOutError));
         }
 
-        if (tcflush(context.fd, TCOFLUSH) != 0)
+        if (tcflush(handle_context.file_descriptor, TCOFLUSH) != 0)
         {
             return cpp_bindings_linux::detail::failErrno<int>(
                 error_callback, cpp_bindings_linux::detail::statusValue(cpp_core::StatusCode::Io::kClearBufferOutError));
