+++ src/serial_get_flow_control.cpp
@@ -10,25 +10,26 @@ extern "C"
 
     MODULE_API auto serialGetFlowControl(int64_t handle, ErrorCallbackT error_callback) -> int
     {
-        cpp_bindings_linux::detail::HandleContext context;
-        const auto rc = cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &context);
-        if (rc < 0)
+        cpp_bindings_linux::detail::HandleContext handle_context;
+        const auto status =
+            cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &handle_context);
+        if (status < 0)
         {
-            return rc;
+            return status;
         }
 
-        termios2 tty{};
-        if (ioctl(context.fd, TCGETS2, &tty) != 0)
+        termios2 serial_settings{};
+        if (ioctl(handle_context.file_descriptor, TCGETS2, &serial_settings) != 0)
         {
             return cpp_bindings_linux::detail::failErrno<int>(
                 error_callback, cpp_bindings_linux::detail::statusValue(cpp_core::StatusCode::Control::kGetStateError));
         }
 
-        if ((tty.c_cflag & CRTSCTS) != 0)
+        if ((serial_settings.c_cflag & CRTSCTS) != 0)
         {
             return 1;
         }
-        if ((tty.c_iflag & (IXON | IXOFF)) != 0)
+        if ((serial_settings.c_iflag & (IXON | IXOFF)) != 0)
         {
             return 2;
         }
