+++ src/serial_get_stop_bits.cpp
@@ -10,21 +10,22 @@ extern "C"
 
     MODULE_API auto serialGetStopBits(int64_t handle, ErrorCallbackT error_callback) -> int
     {
-        cpp_bindings_linux::detail::HandleContext context;
-        const auto rc = cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &context);
-        if (rc < 0)
+        cpp_bindings_linux::detail::HandleContext handle_context;
+        const auto status =
+            cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &handle_context);
+        if (status < 0)
         {
-            return rc;
+            return status;
         }
 
-        termios2 tty{};
-        if (ioctl(context.fd, TCGETS2, &tty) != 0)
+        termios2 serial_settings{};
+        if (ioctl(handle_context.file_descriptor, TCGETS2, &serial_settings) != 0)
         {
             return cpp_bindings_linux::detail::failErrno<int>(
                 error_callback, cpp_bindings_linux::detail::statusValue(cpp_core::StatusCode::Control::kGetStateError));
         }
 
-        return (tty.c_cflag & CSTOPB) != 0 ? 2 : 0;
+        return (serial_settings.c_cflag & CSTOPB) != 0 ? 2 : 0;
     }
 
 } // extern "C"
