+++ src/serial_out_bytes_waiting.cpp
@@ -9,15 +9,16 @@ extern "C"
 
     MODULE_API auto serialOutBytesWaiting(int64_t handle, ErrorCallbackT error_callback) -> int
     {
-        cpp_bindings_linux::detail::HandleContext context;
-        const auto rc = cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &context);
-        if (rc < 0)
+        cpp_bindings_linux::detail::HandleContext handle_context;
+        const auto status =
+            cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &handle_context);
+        if (status < 0)
         {
-            return rc;
+            return status;
         }
 
         int bytes_waiting = 0;
-        if (ioctl(context.fd, TIOCOUTQ, &bytes_waiting) != 0)
+        if (ioctl(handle_context.file_descriptor, TIOCOUTQ, &bytes_waiting) != 0)
         {
             return cpp_bindings_linux::detail::failErrno<int>(
                 error_callback, cpp_bindings_linux::detail::statusValue(cpp_core::StatusCode::Control::kGetStateError));
