+++ src/serial_set_dtr.cpp
@@ -9,15 +9,16 @@ extern "C"
 
     MODULE_API auto serialSetDtr(int64_t handle, int state, ErrorCallbackT error_callback) -> int
     {
-        cpp_bindings_linux::detail::HandleContext context;
-        const auto rc = cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &context);
-        if (rc < 0)
+        cpp_bindings_linux::detail::HandleContext handle_context;
+        const auto status =
+            cpp_bindings_linux::detail::acquireHandleContext<int>(handle, error_callback, &handle_context);
+        if (status < 0)
         {
-            return rc;
+            return status;
         }
 
-        int flag = TIOCM_DTR;
-        if (ioctl(context.fd, state ? TIOCMBIS : TIOCMBIC, &flag) != 0)
+        int modem_flag = TIOCM_DTR;
+        if (ioctl(handle_context.file_descriptor, state ? TIOCMBIS : TIOCMBIC, &modem_flag) != 0)
         {
             return cpp_bindings_linux::detail::failErrno<int>(
                 error_callback, cpp_bindings_linux::detail::statusValue(cpp_core::StatusCode::Control::kSetDtrError));
